create getById Task controller

+++ Back-end/DotNetCore/TMS/TMS/Models/Tasks.cs
@@ -4,7 +4,7 @@ namespace TMS.Models;
 
 public class Tasks
 {
-        private long Id { get; set; }
+        private int Id { get; set; }
         private int storyPoints { get; set; }
         private DateTime dueDate{ get; set; }
         private string username{ get; set; }
